no comment

+++ makeprogs/installprogs.sh
@@ -26,7 +26,7 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE 
 # POSSIBILITY OF SUCH DAMAGE.
 #
-# Last modified 21 February 2003.
+# Last modified 9 March 2003.
 #
 
 #
@@ -34,6 +34,13 @@
 # scripts) into place.  It is not used for Perl library or module (*.pl and
 # *.pm) files.
 #
+
+opt_diff=0
+if [ "$1" = "--diff" ]
+    then
+    opt_diff=1
+    shift
+    fi
 
 USER=$1
 GROUP=$2
@@ -61,17 +68,47 @@ if [ ! -d "$RPM_BUILD_ROOT$DESTDIR" ]
 
 while [ "$1" != "" ]
     do
-    name=`basename "$1"`
-    dest="$DESTDIR/$name"
-    echo "    \"$1\" --> \"$RPM_BUILD_ROOT$dest\""
-    rm -f "$RPM_BUILD_ROOT$dest" || exit 1
-    cp "$1" "$RPM_BUILD_ROOT$dest" || exit 1
-    strip "$RPM_BUILD_ROOT$dest" 2>/dev/null
-    chown $USER "$RPM_BUILD_ROOT$dest" 2>/dev/null
-    chgrp $GROUP "$RPM_BUILD_ROOT$dest" 2>/dev/null
-    chmod $MODE "$RPM_BUILD_ROOT$dest" || exit 1
-    echo "\"$dest\"" >>`dirname $0`/../z_install_begin/installed_files_list
+    file=$1
     shift
+    name=`basename "$file"`
+    dest="$DESTDIR/$name"
+    echo "    \"$file\" --> \"$RPM_BUILD_ROOT$dest\""
+
+    # If this file is to be owned by root, try very hard to to replace
+    # it unless is has really changed.  We do this so that make install
+    # may be run by a non-root user after it has been run once by root.
+    if [ "$USER" = "root" -a -f "$RPM_BUILD_ROOT$dest" -a ! -w "$RPM_BUILD_ROOT$dest" ]
+	then
+	strip $file
+	if diff $file "$RPM_BUILD_ROOT$dest"
+	    then
+	    echo "        (skipping copy because root ownership and unchanged)"
+	    continue
+	    fi
+	fi
+
+    rm -f "$RPM_BUILD_ROOT$dest" || exit 1
+    cp "$file" "$dest" || exit 1
+    strip "$RPM_BUILD_ROOT$dest" 2>/dev/null
+
+    chown $USER "$RPM_BUILD_ROOT$dest" \
+	&& chgrp $GROUP "$RPM_BUILD_ROOT$dest" \
+	&& chmod $MODE "$RPM_BUILD_ROOT$dest"
+    if [ $? -ne 0 ]
+	then
+	if [ "$USER" = "root" ]
+	    then
+	    echo "============================================================================="
+	    echo "The above error almost certainly means that you must re-run make install as"
+	    echo "root (as least in this one directory).  This will happen the first time and"
+	    echo "every time you modify a setuid root program."
+	    echo "============================================================================="
+	    fi
+	exit 1
+	fi
+
+    echo "\"$dest\"" >>`dirname $0`/../z_install_begin/installed_files_list
+
     done
 
 exit 0
